pyproject.toml
[tool.poetry]
name = "metaculus-web"
version = "0.1.0"
description = ""
authors = ["George3d6 <george@cerebralab.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
django = "^5.1.4"
djangorestframework = "^3.15.1"
rest-social-auth = "^8.3.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
python-dateutil = "^2.9.0.post0"
django-extensions = "^3.2.3"
numpy = "^1.26.4"
django-cors-headers = "^4.3.1"
django-anymail = {extras = ["mailgun"], version = "^10.3"}
django-sql-utils = "^0.7.0"
django-dramatiq = "^0.11.6"
django-redis = "^5.4.0"
ipython = "^8.25.0"  # for easier debugging in shell
scipy = "^1.13.1"
django-dynamic-fixture = "^4.0.1"
django-admin-autocomplete-filter = "^0.7.1"
html2text = "^2024.2.26"
gunicorn = "^22.0.0"
django-storages = {extras = ["s3"], version = "^1.14.3"}
pillow = "^10.3.0"
django-imagekit = "^5.0.0"
setuptools = "58.2.0"
jinja2 = "^3.1.5"
apscheduler = "^3.10.4"
openai = "^1.36.0"
pgvector = "^0.3.2"
tiktoken = "^0.7.0"
aiohttp = "^3.10.11"
django-mjml = "^1.2"
markdown = "^3.6"
sshtunnel = "^0.4.0"
python-dotenv = "^1.0.1"
mysql-connector-python = "^9.0.0"
gspread = "^6.1.2"
whitenoise = "^6.7.0"
sentry-sdk = "^2.13.0"
# playwright = "^1.46.0"
brotli = "^1.1.0"
django-better-admin-arrayfield = "^1.4.2"
django-modeltranslation = "^0.19.9"
google-auth = "^2.35.0"
django-select2 = "^8.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-django = "^4.8.0"
pytest-env = "^1.1.3"
factory-boy = "^3.3.0"
django-debug-toolbar = "^4.4.2"
freezegun = "^1.5.1"
pytest-mock = "^3.14.0"
flake8-quotes = "^3.4.0"
playwright = "^1.46.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
